+++ auth.go
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package mysql
+package singlestore
 
 import (
 	"crypto/rand"
@@ -52,7 +52,7 @@ var (
 //	}
 //
 //	if rsaPubKey, ok := pub.(*rsa.PublicKey); ok {
-//		mysql.RegisterServerPubKey("mykey", rsaPubKey)
+//		singlestore.RegisterServerPubKey("mykey", rsaPubKey)
 //	} else {
 //		log.Fatal("not a RSA public key")
 //	}