+++ utils_test.go
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package mysql
+package singlestore
 
 import (
 	"bytes"
@@ -225,7 +225,7 @@ func TestIsolationLevelMapping(t *testing.T) {
 	}
 
 	// check unsupported mapping
-	expectedErr := "mysql: unsupported isolation level: 7"
+	expectedErr := "singlestore: unsupported isolation level: 7"
 	actual, err := mapIsolationLevel(driver.IsolationLevel(sql.LevelLinearizable))
 	if actual != "" || err == nil {
 		t.Fatal("Expected error on unsupported isolation level")